Turn filter functions to factories
To avoid having to pass arguments in query function, it is done when
initializing filter. The reason why this is needed is because there will
be clock filters and clock grouping which will also be passed to the
query and passing arguments for each filter in the query quickly becomes unwieldy.